fix: fix review feedbacks
