+++ packages/screens/Marketplace/types.ts
@@ -6,3 +6,5 @@ export type NFTActivityType = {
   buyer: string;
   seller: string;
 };
+
+export type TabsListType = "allNFTs" | "owned" | "activity";
