+++ packages/screens/RiotGame/RiotGameEnrollScreen.tsx
@@ -36,7 +36,7 @@ import {
 import { layout } from "../../utils/style/layout";
 import { EnrollSlot } from "./component/EnrollSlot";
 import { GameContentView } from "./component/GameContentView";
-import { RipperSelectorModal } from "./component/RipperSelectorModalV2";
+import { RipperSelectorModal } from "./component/RipperGridSelectorModal";
 import { SimpleButton } from "./component/SimpleButton";
 
 const RIPPER_SLOTS = [0, 1, 2, 3, 4, 5];
